Add doc for WUC (Word Under Cursor) feature.
Add pragmatic extension to tag-lookup to adapt to recent change to WUC,
which extended the WUC to include words to the left joined by
hljoinchars-defined charset.  Adapatation is to second-chance-try the
rightmost word-only segment of the WUC string (thus effectively trying the
original-definition WUC only after new-WUC lookup has failed).