fix mfreplace on lines containing tabs in not-previously focused FBUFs
Issue was that during mfreplace, FBUFs which have not yet received focus in
this editor session may be "hit".  The act of receiving focus (which will
only occur if the "hit" needs to be interactively approved by the user) can,
if happening for the first time, alter the tabwidth setting of the FBUF
(which is used to provide the preview-display), but the search (for at least
the first "hit") will have taken place with a default tabwidth setting.