improve timing of reading of <tagged-files>
See last para below for related design discussion!

Lua:
  Existing (Lua) code that writes content of this pseudofile now called by
  ...existing but heretofore unused hook ReadPseudoFileOk_FROM_C().
  Removed now unused single-purpose C++ -> Lua API readtagsfile.

C++:
  fExecute("readtagsfile") replaced with 'arg "<tagged-files>" setfile'
  This is clunky as it has visual side-effects (focus is momentarily put
  on "<tagged-files>" pseudofile (iff it is actually written)).  Leaving
  in as (1) I'm lazy, (2) while clunky, the visual glitch provides user
  feedback (the user obviously needs to understand what's happening... in
  order to make use of this feedback, which is why it remains clunky.

This change also fixes the problem of "sb" command (or any other way of)
switching to the "<tagged-files>" pseudofile, leading to the inappropriate
displaying of an empty version of "<tagged-files>"; overall, the filling
of the "<tagged-files>" pseudofile is now automatic upon reference just
like other "system [f]bufs".

Design Discussion: I've had a longstanding "internal argument"
with myself whether it would be approproate to refresh "<tagged-files>"
each time it gets focus (which is, I'm pretty sure, how it works after
this commit) but for the time being, this behavior is better than the
prior, more magical behaviors.  Also, the longstanding dilemma regarding
auto-refreshing sysbufs is that any user edits are unilaterally
obliterated.  This larger question (whether to allow user edits of such
pseudofiles, and if so, whether/how to preserve user edits) is one that
I might need to come up with a better answer for.