display subprocess cmdline rather `shellnm tmpfnm`
Both CMD and bash child processes have (on Win32) long been written
to a temp scriptfile and CMD/bash called to execute the temp scriptfile.

From a user-record perspective, this can be problematic because the
actual cmdline executed, often constructed on the fly in e.g. <win0>
might depending on the autorefresh nature of the FBUF in which the
cmdline was constructed, by overwritten by FBUF autorefresh as soon
as the cmdline was staged for execution (which means the current FBUF
is switched away from, and any switch back to it will result in it
being autorefreshed.

So rather than being literal and displaying a cmdline that looks like
e.g. `bash temp/scriptname.sh`, the cmdline which
e.g. temp/scriptname.sh _contains_ is displayed.