Use -MMD to avoid pre-gen of .makedeps files during clean build
By moving generation of .makedeps files into .o-file generation (a.k.a.
compile) phase.

Note: a cmdtbl.dat change will force rebuild of not just the
makedeps-analysis-based set of .o files but rather ALL .o files.
The difference between these two sets of .o files is negligible, and
cmdtbl.dat changes so infrequently that this is effectively a don't care.